Partially implemented code restructure
