+++ data/VERSION
@@ -1,0 +1,1 @@
+2.0.0
