+++ player/utils/sys_logger.py
@@ -1,19 +1,18 @@
 import logging
+import sys
 
 def setup_logging():
-    logging.basicConfig(level=logging.INFO,
+    logging.basicConfig(level=logging.DEBUG,
                         datefmt="%m-%d %H:%M",
                         filename="output.log",
                         filemode="w")
 
-    console = logging.StreamHandler()
-    console.setLevel(logging.INFO)
-
     formatter = logging.Formatter("%(name)-12s: %(levelname)-8s %(message)s")
 
+    console = logging.StreamHandler(sys.stdout)
+    console.setLevel(logging.DEBUG)
     console.setFormatter(formatter)
 
-    logging.getLogger("").addHandler(console)
-
+    logging.getLogger("app.stdout").addHandler(console)
     logging.getLogger("app.logger").info("Logger has been initialized")
