+++ player/widgets/__init__.py
@@ -1,0 +1,2 @@
+from player.widgets.hover_button import HoverButton
+from player.widgets.scrollbar_frame import ScrollbarFrame
