+++ player/window/main.py
@@ -8,15 +8,16 @@ from tkinter import *
 from tkinter import messagebox
 from tkinter.font import Font
 
-import player.window.widgets as widgets
-import player.data as data
-import player.audio as audio
-import player.files as files
-import player.timer as timer
-import player.download as download
-import player.settings as settings
-import player.updater as updater
-import player.playlist as playlist
+import player.widgets as widgets
+import player.utils.data as data
+import player.utils.audio as audio
+import player.utils.timer as timer
+import player.utils.constants as constants
+import player.menus.files as files
+import player.menus.download as download
+import player.menus.settings as settings
+import player.menus.updater as updater
+import player.menus.playlists as playlist
 
 BACK_COLOUR = data.view("back_colour", "c")
 FORE_COLOUR = data.view("fore_colour", "c")
@@ -32,14 +33,9 @@ class MainWindow(Tk):
         self.wm_title("Music Player")
         self.geometry("850x600")
         self.resizable(False, False)
-
-        self.assets = {
-            "pauseplay": PhotoImage(file="player/Assets/pauseplay.png"),
-            "stop": PhotoImage(file="player/Assets/stopicon.png"),
-            "cascadia": Font(size=10, family="Cascadia Mono"),
-            "small_cascadia": Font(size=8, family="Cascadia Mono"),
-            # "loop": PhotoImage(file="player/Assets/loopicon.png"),
-        }
+        
+        self.fonts = constants.Font()
+        self.images = constants.Image()
 
         self.sbf = None
         self.current_song = None
@@ -49,7 +45,7 @@ class MainWindow(Tk):
         self.bind("<Escape>", self.close_window)
 
         try:
-            self.iconbitmap("player/Assets/musical_note.ico")
+            self.iconbitmap(self.images.MAIN)
         except TclError:
             pass
 
@@ -69,40 +65,40 @@ class MainWindow(Tk):
         Frame(self, bg=data.view("accent_colour", "c"), height=2, bd=0).pack(fill=X, side=BOTTOM, pady=100)
         Frame(self, bg=data.view("accent_colour", "c"), width=2, height=500, bd=0 ).place(x=198, y=0)
 
-        self.duration_label = Label(self, text="00:00 / 00:00", bg=FORE_COLOUR, fg="white", font=self.assets["cascadia"])
+        self.duration_label = Label(self, text="00:00 / 00:00", bg=FORE_COLOUR, fg="white", font=self.fonts.MAIN)
         self.duration_label.place(x=250, y=545)
         
-        self.now_playing = Label(self, text="Nothing is playing. Play a song by finding a song in the song list and clicking the ▶ next to it!", bg=FORE_COLOUR, fg="white", font=self.assets["cascadia"])
+        self.now_playing = Label(self, text="Nothing is playing. Play a song by finding a song in the song list and clicking the ▶ next to it!", bg=FORE_COLOUR, fg="white", font=self.fonts.MAIN)
         self.now_playing.place(relx=0.5, rely=0.86, anchor=CENTER)
         
-        self.ispaused = Label(self, text="", bg=FORE_COLOUR, fg="white", font=self.assets["cascadia"])
+        self.ispaused = Label(self, text="", bg=FORE_COLOUR, fg="white", font=self.fonts.MAIN)
         self.ispaused.place(relx=0.5, rely=0.97, anchor=CENTER)
 
-        self.pauseplay_button = Button(self, image=self.assets["pauseplay"], background=FORE_COLOUR, borderwidth=0, command=self.pause_or_resume, activebackground=FORE_COLOUR)
+        self.pauseplay_button = Button(self, image=self.images.PAUSEPLAY, background=FORE_COLOUR, borderwidth=0, command=self.pause_or_resume, activebackground=FORE_COLOUR)
         self.pauseplay_button.place(relx=0.499, rely=0.93, anchor=CENTER)
 
         self.addmusic_button = widgets.HoverButton(self, text="Add music", bg=BACK_COLOUR, fg="white", compound="left", borderwidth=0, command=self.download.download_window, activebackground=BACK_COLOUR, activeforeground=data.view("accent_colour", "c"))
-        self.addmusic_button["font"] = Font(size=12, family="Cascadia Mono", weight="bold")
+        self.addmusic_button["font"] = self.fonts.MEDIUM
         self.addmusic_button.place(x=25, y=40)
 
         self.rename_file = widgets.HoverButton(self, text="Rename a file", bg=BACK_COLOUR, fg="white", compound="left", borderwidth=0, command=self.files.rename_window, activebackground=BACK_COLOUR, activeforeground=data.view("accent_colour", "c"))
-        self.rename_file["font"] = Font(size=12, family="Cascadia Mono", weight="bold")
+        self.rename_file["font"] = self.fonts.MEDIUM
         self.rename_file.place(x=25, y=70)
         
         self.settings_btn = widgets.HoverButton(self, text="Settings", bg=BACK_COLOUR, fg="white", compound="left", borderwidth=0, command=self.settings.settings_window, activebackground=BACK_COLOUR, activeforeground=data.view("accent_colour", "c"))
-        self.settings_btn["font"] = Font(size=12, family="Cascadia Mono", weight="bold")
+        self.settings_btn["font"] = self.fonts.MEDIUM
         self.settings_btn.place(x=25, y=100)
 
         self.delete_song = widgets.HoverButton(self, text="Delete a file", bg=BACK_COLOUR, fg="white", compound="left", borderwidth=0, command=self.files.delete_file, activebackground=BACK_COLOUR, activeforeground=data.view("accent_colour", "c"))
-        self.delete_song["font"] = Font(size=12, family="Cascadia Mono", weight="bold")
+        self.delete_song["font"] = self.fonts.MEDIUM
         self.delete_song.place(x=25, y=130)
 
         self.myplaylists = widgets.HoverButton(self, text="My Playlists", bg=BACK_COLOUR, fg="white", compound="left", borderwidth=0, command=playlist.show_playlists, activebackground=BACK_COLOUR, activeforeground=data.view("accent_colour", "c"))
-        self.myplaylists["font"] = Font(size=12, family="Cascadia Mono", weight="bold")
+        self.myplaylists["font"] = self.fonts.MEDIUM
         self.myplaylists.place(x=25, y=160)
 
         self.update = widgets.HoverButton(self, text="Check for updates", bg=BACK_COLOUR, fg="white", compound="left", borderwidth=0, command=updater.check_updates, activebackground=BACK_COLOUR, activeforeground=data.view("accent_colour", "c"))
-        self.update["font"] = Font(size=12, family="Cascadia Mono", weight="bold")
+        self.update["font"] = self.fonts.MEDIUM
         self.update.place(x=15, y=460)
 
         self.volume = Scale(self, orient=HORIZONTAL, variable=DoubleVar(), bg=FORE_COLOUR, fg="white", troughcolor=BACK_COLOUR, highlightthickness=0)
@@ -110,7 +106,7 @@ class MainWindow(Tk):
         self.volume.bind("<ButtonRelease-1>", self.set_volume)
         self.volume.place(x=530, y=535)
 
-        self.stop_button = Button(self, image=self.assets["stop"], background=FORE_COLOUR, borderwidth=0, command=self.stop, activebackground=FORE_COLOUR)
+        self.stop_button = Button(self, image=self.images.STOP, background=FORE_COLOUR, borderwidth=0, command=self.stop, activebackground=FORE_COLOUR)
         self.stop_button.place(x=375, y=545)
 
     def set_np(self, text: str):
@@ -136,7 +132,7 @@ class MainWindow(Tk):
                 text="Songs that you have will appear here, but you don't have any!",
                 bg=data.view("songlist_colour", "c"),
                 fg="white",
-                font=self.assets["cascadia"]
+                font=self.fonts.MAIN
             ).grid(row=0, column=0)
             
             refnosongs = Button(
@@ -144,7 +140,7 @@ class MainWindow(Tk):
                 text="[Refresh]",
                 bg=data.view("songlist_colour", "c"),
                 fg="white",
-                font=self.assets["cascadia"],
+                font=self.fonts.MAIN,
                 command=self.refresh_songlist,
                 borderwidth=0
             )
@@ -156,7 +152,7 @@ class MainWindow(Tk):
             text="[Refresh]",
             bg=data.view("songlist_colour", "c"),
             fg="white",
-            font=self.assets["cascadia"],
+            font=self.fonts.MAIN,
             borderwidth=0,
             height=2,
             width=10,
@@ -198,7 +194,7 @@ class MainWindow(Tk):
                 self.scroll_frame,
                 text=file,
                 bg=data.view("songlist_colour", "c"),
-                font=self.assets["cascadia"],
+                font=self.fonts.MAIN,
                 fg="white",
             ).grid(row=i, column=1, sticky="w")
 
@@ -206,7 +202,7 @@ class MainWindow(Tk):
                 self.scroll_frame,
                 text="        " + duration,
                 bg=data.view("songlist_colour", "c"),
-                font=self.assets["cascadia"],
+                font=self.fonts.MAIN,
                 fg="white"
             ).grid(row=i, column=2, sticky="e")
